refactor(Pagination): inline `PaginationLink`
- Inline `PaginationLink` in `Pagination.tsx`
- Forward `createClickHandler()`
